requirements.txt
# Ambiente e variáveis
python-dotenv==1.0.1

# Requisições e integração com APIs
requests==2.32.3

# Interface de linha de comando (se ainda usada)
click==8.1.7

# Interface visual web
streamlit

# Estilização no terminal (usado no início do projeto)
rich==13.9.2

# Processamento e exportação de dados
pandas==2.2.2

# ORM para banco de dados
SQLAlchemy==2.0.35

# Acesso a PostgreSQL
psycopg2-binary==2.9.9
psycopg[binary]==3.2.1  # Incluído conforme solicitado (mas escolha um dos dois se necessário)

# Login seguro (hash de senha)
passlib==1.7.4

# Gemini AI
google-generativeai==0.3.2

# Testes e cobertura
pytest==8.2.1
pytest-cov==5.0.0
